Update 1/10/24
Added basic filtering option '-o' to only show/export open ports.

If an open port is detected, it will try to grab the service banner for that port.

For some reason the conditional at the end of the program will default to the 3rd option even when 'e' or 'v' is given. Will need to resolve this later.